dash/requirements.txt
dash==2.6.1
pandas==1.4.3
snowflake-sqlalchemy==1.3.4